Added pdfgen.html file
This file is a new page for the website with a button back to the home page, some text, an embedded image, some formatting, a button to download a pdf page, and functionality to the button that downloads a pdf. The downloaded pdf has a qr code in the bottom right corner.

* Added html file
* Added a button
* Added functionality to the button
* Added text with brief descriptions of what's happening on the page

addresses #14