Style and Fix: Made Template into proper page.
The template page now has the Nav bar and looks consistent with the other pages. The nav bar also has a proper link to the template
page and the template page can go to all the other pages. This helps make it look better and makes it part of the whole website.

* Fix: Template page is now connected and accessable on website.
* Nav bar added to Template page.
* Nav bar has link to Template on all pages.
* Template page and features more consistent to look of the rest of the website.